Create README.md
This program takes a sudoku puzzle as an input and gives the solutions as an output.  The format is weird.  You need to put all the values for the puzzle on one line with dots where the blanks should be.  Write this input in a file and redirect that file as input to the program.